+++ CHANGELOG.md
@@ -11,7 +11,9 @@
   - fix(Autocompleter): check if its mounted before modifying state
   - feat(tooltip): Initial creation / promotion of tooltip component, intended to display short help or reminder text
   - feat(help-icon): Initial creation HelpIcon component
-  - feat(input): expose `tooltip` prop which renders a help icon next to the label if provided
+  - Added help text icon as optional prop the following components:
+    - input
+    - abstract-custom-field  
 </details>
 
 ## v0.75.0